fix(geodesicfitting): fix orbital elements computation and Newtonian conversion procedure
Fix the orbital elements computation procedure by using the inverse of the
parametric functions that compute the synthetic data to properly compute
the orbital elements. Also, the conversion from the Newtonian
parametrization to the first integrals of motion was fixed by removing
the different branches of energies because it was incorrect.